+++ packages/elements/empty-state/examples/1-only-mandatory-props.js
@@ -1,0 +1,10 @@
+// @flow
+
+import React from 'react';
+import EmptyState from '../src/EmptyState';
+
+const props = {
+  header: 'I am the header',
+};
+
+export default () => <EmptyState {...props} />;